2nd commit

+++ src/main/java/com/sd/thrillio/controllers/BookmarkController.java
@@ -27,48 +27,53 @@ public class BookmarkController extends HttpServlet{
 	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		long userId = 5;
 		RequestDispatcher despatcher = null;
 		
-		if(request.getServletPath().contains("save")) {
-			//save book.
-			String bookId = request.getParameter("bid");
-			User user = UserManager.getInstance().getUser(userId);
-			Bookmark bookmark = BookmarkManager.getInstance().getBook(Long.parseLong(bookId));
+		if(request.getSession().getAttribute("userId")!=null) {
 			
-			BookmarkManager.getInstance().saveUserBookmark(user, bookmark);
+			long userId = (long) request.getSession().getAttribute("userId");
 			
-			Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(true, userId);
-			request.setAttribute("books", list);
-			despatcher = request.getRequestDispatcher("/mybookmark.jsp");
-		}
-		else if(request.getServletPath().contains("mybooks")) {
-			Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(true, userId);
-			request.setAttribute("books", list);
-			despatcher = request.getRequestDispatcher("/mybookmark.jsp");
-		}
-		else if(request.getServletPath().contains("delete")) {
-			//delete book
-			String bookId = request.getParameter("bid");
-			User user = UserManager.getInstance().getUser(userId);
-			Bookmark bookmark = BookmarkManager.getInstance().getBook(Long.parseLong(bookId));
-			
-			BookmarkManager.getInstance().deleteUserBookmark(user, bookmark);
-			
-			Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(true, userId);
-			request.setAttribute("books", list);
-			despatcher = request.getRequestDispatcher("/mybookmark.jsp");
-			
+			if(request.getServletPath().contains("save")) {
+				//save book.
+				String bookId = request.getParameter("bid");
+				User user = UserManager.getInstance().getUser(userId);
+				Bookmark bookmark = BookmarkManager.getInstance().getBook(Long.parseLong(bookId));
+				
+				BookmarkManager.getInstance().saveUserBookmark(user, bookmark);
+				
+				Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(true, userId);
+				request.setAttribute("books", list);
+				despatcher = request.getRequestDispatcher("/mybookmark.jsp");
+			}
+			else if(request.getServletPath().contains("mybooks")) {
+				Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(true, userId);
+				request.setAttribute("books", list);
+				despatcher = request.getRequestDispatcher("/mybookmark.jsp");
+			}
+			else if(request.getServletPath().contains("delete")) {
+				//delete book
+				String bookId = request.getParameter("bid");
+				User user = UserManager.getInstance().getUser(userId);
+				Bookmark bookmark = BookmarkManager.getInstance().getBook(Long.parseLong(bookId));
+				
+				BookmarkManager.getInstance().deleteUserBookmark(user, bookmark);
+				
+				Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(true, userId);
+				request.setAttribute("books", list);
+				despatcher = request.getRequestDispatcher("/mybookmark.jsp");
+				
+			}
+			else {
+				Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(false, userId);
+				request.setAttribute("books", list);
+				despatcher = request.getRequestDispatcher("/browse.jsp");
+			}
 		}
 		else {
-			Collection<Bookmark> list = BookmarkManager.getInstance().getBooks(false, userId);
-			request.setAttribute("books", list);
-			despatcher = request.getRequestDispatcher("/browse.jsp");
+			despatcher = request.getRequestDispatcher("/login.jsp");
 		}
-		despatcher.forward(request, response);
 		
-		
-		
+		despatcher.forward(request, response);		
 	}	
 
 	/**
